add key per auditLogCard + combine FilterActivityTypes to AuditLog.js page
