+++ src/Pages/AuditLog/AuditLog.js
@@ -3,7 +3,6 @@ import { getAllLogs } from '../../APIFunctions/AuditLog';
 import Pagination from './Components/Pagination';
 import { useUser } from '../../Components/context/UserContext';
 import AuditLogCard from './Components/AuditLogCard';
-import FilterActivityTypes from './Components/FilterActivityTypes';
 
 export default function AuditLogPage() {
   const [auditLogsData, setAuditLogsData] = useState({ items: [], totalLogs: 0 });
@@ -11,6 +10,7 @@ export default function AuditLogPage() {
   // states for filters
   const [searchQuery, setSearchQuery] = useState('');
   const [activityFilters, setActivityFilters] = useState([]);
+  const [isDropdownOpen, setIsDropdownOpen] = useState(false);
 
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState('');
@@ -18,6 +18,21 @@ export default function AuditLogPage() {
   const [applyingFilters, setApplyingFilters] = useState(false);
 
   const user = useUser();
+
+  const toggleActivityFilter = activity => {
+    setActivityFilters(prev => (prev.includes(activity) ? prev.filter(a => a !== activity) : [...prev, activity]));
+  };
+
+  const activityTypes = [
+    'SIGN_UP',
+    'LOG_IN',
+    'UPDATE_USER',
+    'PRINT_PAGE',
+    'VERIFY_EMAIL',
+    'EMAIL_SENT',
+    'CHANGE_PW',
+    'RESET_PW',
+  ];
 
   const getAuditLogsFromDB = async () => {
     try {
@@ -138,7 +153,46 @@ export default function AuditLogPage() {
                 }}
               />
             </div>
-            <FilterActivityTypes activityFilters={activityFilters} setActivityFilters={setActivityFilters} />
+            <div className='relative'>
+              <label className='block text-sm font-medium text-gray-300 mb-2'>Filter by Activity Type</label>
+              <button
+                onClick={() => setIsDropdownOpen(!isDropdownOpen)}
+                className='w-full px-3 py-2 bg-gray-700 border border-gray-600 rounded-md text-white text-left focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500 flex justify-between items-center'
+              >
+                <span>
+                  {activityFilters.length === 0 ? 'Select activities...' : `${activityFilters.length} selected`}
+                </span>
+                <svg
+                  className={`w-4 h-4 transition-transform ${isDropdownOpen ? 'rotate-180' : ''}`}
+                  fill='none'
+                  stroke='currentColor'
+                  viewBox='0 0 24 24'
+                >
+                  <path strokeLinecap='round' strokeLinejoin='round' strokeWidth={2} d='M19 9l-7 7-7-7' />
+                </svg>
+              </button>
+
+              {isDropdownOpen && (
+                <div className='absolute z-10 w-full mt-1 bg-gray-700 border border-gray-600 rounded-md shadow-lg max-h-60 overflow-y-auto'>
+                  <div className='p-2'>
+                    {activityTypes.map(activity => (
+                      <label
+                        key={activity}
+                        className='flex items-center space-x-2 p-2 hover:bg-gray-600 rounded cursor-pointer'
+                      >
+                        <input
+                          type='checkbox'
+                          checked={activityFilters.includes(activity)}
+                          onChange={() => toggleActivityFilter(activity)}
+                          className='form-checkbox h-4 w-4 text-blue-600 bg-gray-700 border-gray-600 rounded focus:ring-blue-500'
+                        />
+                        <span className='text-white text-sm'>{activity.replace(/_/g, ' ')}</span>
+                      </label>
+                    ))}
+                  </div>
+                </div>
+              )}
+            </div>
             <div className='flex gap-2'>
               <button
                 onClick={applyFilters}
@@ -156,7 +210,7 @@ export default function AuditLogPage() {
           </div>
         </div>
         {auditLogsData.items.map((log, index) => (
-          <AuditLogCard log={log} index={index} />
+          <AuditLogCard key={log._id || index} log={log} index={index} />
         ))}
       </div>
       {maybeRenderAuditLogs()}
