+++ src/Pages/AuditLog/Components/AuditLogCard.js
@@ -48,7 +48,7 @@ const AuditLogCard = ({ log, index }) => {
     // checks if a user updates or deletes ANOTHER user
     if (action === 'UPDATE_USER') {
       if (log.documentId && log.documentId !== log.userId) {
-        return 'updated another user\'s account information';
+        return "updated another user's account information";
       }
       return 'updated their account information';
     }
@@ -68,10 +68,7 @@ const AuditLogCard = ({ log, index }) => {
   };
 
   return (
-    <div
-      key={log._id || index}
-      className='p-6 rounded-lg shadow-sm transition-all duration-200 hover:shadow-md bg-gray-800 border border-gray-700 hover:bg-gray-750'
-    >
+    <div className='p-6 rounded-lg shadow-sm transition-all duration-200 hover:shadow-md bg-gray-800 border border-gray-700 hover:bg-gray-750'>
       <div className='flex items-start justify-between'>
         <div className='flex-1'>
           <div className='flex items-center space-x-3 mb-2'>
